requirements.txt
keras==2.12.0
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
scikit_learn==1.2.2
